Add components, enums, utils folders as src subfolder - organize files accordingly. Update import statements accordingly. Remove unused logo.svg asset. Replace color.controller.js magic number with descriptive variable.
