+++ client/src/components/ColorDetailView.css
