+++ client/src/enums/ViewConfiguration.js
